refactor: 商品のルータ作成
機能や責務ごとにエンドポイントをグループ化して別ファイルとして整理
