+++ src/FastAPI/fastapi-introduction/README.md
@@ -75,3 +75,14 @@ curl localhost:8000
 http://localhost:8000/docsにアクセスします。
 APIの更新があると自動的に内容は更新されます。
 http://localhost:8000/docs画面でAPIの動作確認を行うことができます。
+
+
+## ルータ
+- webアプリやAPIにおけるエンドポイントの管理を助けるツール
+- ユーザーがアクセスするURLと、それに対応する処理を関連づける役割
+- 対応する処理を関連づけることをルーティングと呼びます
+
+### ルーターを使用するメリット
+・機能や責務ごとにエンドポイントをグループ化して別ファイルとして整理することができる
+・エンドポイントの再利用が可能になる
+・異なるモジュールやアプリケーションの部分を分離して開発可能になる
